♻️Refactor: Extract class from PlayerController creating PlayerWeapon class
