+++ Assets/Prefabs/Player.prefab
@@ -95,6 +95,7 @@ GameObject:
   - component: {fileID: 8433152684776134600}
   - component: {fileID: 8433152684776134607}
   - component: {fileID: 8433152684776134601}
+  - component: {fileID: 7907779579777558085}
   - component: {fileID: 8433152684776134603}
   - component: {fileID: 8433152684776134602}
   m_Layer: 0
@@ -186,19 +187,28 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1e8e286bfa99d204f9bd1ae613550f8f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  bulletPrefab: {fileID: 4221975376070287205, guid: 1f022f2f44de41e44aacce5411ca0126, type: 3}
-  shootTransform: {fileID: 8433152685066387617}
   speed: 100
-  rotation: 0
-  shootCoolDown: 0.4
+--- !u!114 &7907779579777558085
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8433152684776134606}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: eeb722044656d77429bded63ab9937f5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   primaryWeaponSpawnPoints:
   - {fileID: 8433152685066387617}
   secondaryWeaponSpawnPoints:
-  - {fileID: 8433152685066387617}
   - {fileID: 8433152684230825743}
+  - {fileID: 8433152685066387617}
   - {fileID: 8433152685071985261}
   singleShotBulet: {fileID: 11400000, guid: f5226c61f27138f4a94bcaa39a8cbfa4, type: 2}
   tripleShotBulet: {fileID: 11400000, guid: 54196ae79ad71714f93be35a3408e2b4, type: 2}
+  bulletPrefab: {fileID: 4221975376070287205, guid: 1f022f2f44de41e44aacce5411ca0126, type: 3}
 --- !u!50 &8433152684776134603
 Rigidbody2D:
   serializedVersion: 4
