+++ Assets/Scripts/Player/PlayerController.cs
@@ -4,18 +4,8 @@ using UnityEngine;
 
 public class PlayerController : MonoBehaviour
 {
-    public Rigidbody2D bulletPrefab;
-    public Transform shootTransform;
-
     [SerializeField] private float speed;
 
-    [SerializeField] private List<Transform> primaryWeaponSpawnPoints = new List<Transform>();
-    [SerializeField] private List<Transform> secondaryWeaponSpawnPoints = new List<Transform>();
-
-    public BulletSO singleShotBulet;
-    public BulletSO tripleShotBulet;
-
-    private bool _canShoot = true;
     private Rigidbody2D playerRigidbody;
 
     void Start()
@@ -25,8 +15,6 @@ public class PlayerController : MonoBehaviour
 
     void Update()
     {
-        //SingleShoot();
-        TripleShoot();
         PlayerMovement();
     }
 
@@ -81,36 +69,4 @@ public class PlayerController : MonoBehaviour
 
     }
 
-    private void SingleShoot()
-    {
-        Shoot(primaryWeaponSpawnPoints, singleShotBulet);
-    }
-
-    private void TripleShoot()
-    {
-        Shoot(secondaryWeaponSpawnPoints, tripleShotBulet);
-    }
-
-    private void Shoot(List<Transform> spawnPoints, BulletSO bulletSo)
-    {
-        if (_canShoot)
-        {
-            foreach (var spawnPoint in spawnPoints)
-            {
-                var rotation = (transform.localEulerAngles.z - 90) * Mathf.PI / 180;
-                var bulletDirection = new Vector2(Mathf.Cos(rotation), Mathf.Sin(rotation));
-
-                Rigidbody2D bulletRigidbody = Instantiate(bulletPrefab, spawnPoint.position, Quaternion.identity);
-                bulletRigidbody.velocity = -bulletSo.Speed * Time.fixedDeltaTime * bulletDirection;
-            }
-            StartCoroutine(ShootCoolDown(bulletSo.CoolDown));
-        }
-    }
-
-    private IEnumerator ShootCoolDown(float shootCoolDown)
-    {
-        _canShoot = false;
-        yield return new WaitForSeconds(shootCoolDown);
-        _canShoot = true;
-    }
 }
