+++ Assets/Scripts/Player/PlayerWeapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eeb722044656d77429bded63ab9937f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
