+++ Assets/Scripts/Weapon.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7a13ea7a105476d4da2ea94a74faa039
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
